+++ src/components/form/form.js
@@ -112,13 +112,12 @@ class Form extends React.Component {
     cancelText: React.PropTypes.string,
 
     /**
-     * Gives the cancel button a color.
+     * Properties for the cancel button
      *
-     * @property cancelTheme
-     * @type {String}
-     * @default blue
+     * @property cancelButtonProps
+     * @type {Object}
      */
-    cancelTheme: React.PropTypes.string,
+    cancelButtonProps: React.PropTypes.object,
 
     /**
      * Text for the save button
@@ -130,13 +129,12 @@ class Form extends React.Component {
     saveText: React.PropTypes.string,
 
     /**
-     * Gives the save button a color.
+     * Properties for the save button
      *
-     * @property saveTheme
-     * @type {String}
-     * @default blue
+     * @property saveButtonProps
+     * @type {Object}
      */
-    saveTheme: React.PropTypes.string,
+    saveButtonProps: React.PropTypes.object,
 
     /**
      * Custom function for Cancel button onClick
@@ -176,8 +174,6 @@ class Form extends React.Component {
     buttonAlign: 'right',
     cancel: true,
     save: true,
-    saveTheme: 'blue',
-    cancelTheme: 'blue',
     saving: false,
     validateOnMount: false
   }
@@ -486,10 +482,11 @@ class Form extends React.Component {
    * @return {Object} JSX cancel button
    */
   get cancelButton() {
-    let cancelClasses = "carbon-form__cancel";
+    let cancelClasses = "carbon-form__cancel",
+        cancelProps = Object.assign({}, this.props.cancelButtonProps, { type: 'button', onClick: this.cancelForm });
 
     return (<div className={ cancelClasses }>
-      <Button type='button' theme={ this.props.cancelTheme } onClick={ this.cancelForm } >
+      <Button {...cancelProps}>
         { this.props.cancelText || I18n.t('actions.cancel', { defaultValue: 'Cancel' }) }
       </Button>
     </div>);
@@ -511,13 +508,13 @@ class Form extends React.Component {
    * @return {Object} JSX save button
    */
   get saveButton() {
-    let errorCount;
-
-    let saveClasses = classNames(
-      "carbon-form__save", {
-        "carbon-form__save--invalid": this.state.errorCount || this.state.warningCount
-      }
-    );
+    let errorCount,
+        saveClasses = classNames(
+          "carbon-form__save", {
+            "carbon-form__save--invalid": this.state.errorCount || this.state.warningCount
+          }
+        ),
+        saveProps = Object.assign({}, this.props.saveButtonProps, { as: 'primary', disabled: this.props.saving });
 
     if (this.state.errorCount || this.state.warningCount) {
       // set error message (allow for HTML in the message - https://facebook.github.io/react/tips/dangerously-set-inner-html.html)
@@ -532,7 +529,7 @@ class Form extends React.Component {
     return (
       <div className={ saveClasses }>
         { errorCount }
-        <Button as="primary" theme={ this.props.saveTheme } disabled={ this.props.saving }>
+        <Button { ...saveProps }>
           { this.props.saveText || I18n.t('actions.save', { defaultValue: 'Save' }) }
         </Button>
       </div>